requirements.txt
streamlit==0.78.0
Keras==2.4.3
h5py==2.10.0
opencv_contrib_python==4.5.1.48
twilio==6.53.0
fpdf==1.7.2
numpy==1.19.5
Pillow==8.1.2
progressbar33==2.4
python-dotenv==0.15.0
